requirements.txt
pandas
camelot-py
